move agent components to subdir
