+++ conversational-client/src/app/app.component.html
@@ -14,7 +14,7 @@
       <app-history-tutor></app-history-tutor>
     </div>
     <div *ngIf="agentId === 'fake'" class="prompt-mode-container">
-      <app-fake-mode></app-fake-mode>
+      <app-fake-agent></app-fake-agent>
     </div>
     <!-- End custom UI -->
