+++ conversational-client/src/app/components/agents/fake-agent/fake-agent.component.ts
@@ -2,12 +2,13 @@ import {Component} from '@angular/core';
 import {NgZone} from '@angular/core';
 
 @Component({
-  selector: 'app-fake-mode',
-  templateUrl: './fake-mode.component.html',
-  styleUrl: './fake-mode.component.css',
+  selector: 'app-fake-agent',
+  templateUrl: './fake-agent.component.html',
+  styleUrl: './fake-agent.component.css',
 })
-export class FakeModeComponent {
+export class FakeAgentComponent {
   worldStateDisplay: string = '';
+
   constructor(private zone: NgZone) {}
 
   updateWorldState(worldState: object): object {
